+++ utils/RCBot2_meta/bot_ga.h
@@ -66,7 +66,7 @@ public:
 
 	void setGA ( CGA *ga ) { m_ga = ga; }
 	// size of population
-	unsigned int size () const { return m_theIndividuals.size(); }
+	size_t size () const { return m_theIndividuals.size(); }
 
 	// get from population index
 	IIndividual *get ( int iIndex ) const;
